Arregla los 500 de la API y endurece secretos, puertos y límites
Críticos:
- #3 penalizeMatch: responded_at = CASE ... THEN \$3::timestamptz. Sin el cast
  Postgres falla con 42804, synchronizeDueAuctions revienta y toda la API queda
  caída de forma permanente.
- #4/#5/#6 ledger_entries: -\$2::bigint / \$2::bigint en BID_HOLD, BID_RELEASE y
  WITHDRAWAL. El menos unario sobre un parámetro sin tipo da 42725.
- #1/#2/#31 secretos: rechaza en cualquier entorno todo valor publicado en el repo,
  exige largo mínimo, ADMIN_PASSWORD obligatorio sin valor por defecto y
  DEMO_PASSWORD opcional (sin él, el seed demo no crea cuentas).
- #7 puertos: postgres, backend y frontend se publican sólo en loopback por defecto.

Medios:
- #26 límite estricto y propio en /auth/login (5/min por IP y por correo) y en
  /auth/register, además del tope global.
- #28 trust proxy siempre, configurable con TRUST_PROXY_HOPS.
- #35 /health informa el estado de la base por separado, con su código real, y
  responde 503 en vez del 500 genérico.

+++ backend/src/routes/auctions.js
@@ -270,7 +270,7 @@ router.post(
         `INSERT INTO ledger_entries
           (user_id, entry_type, available_delta, frozen_delta,
            auction_id, bid_id, description)
-         VALUES ($1, 'BID_HOLD', -$2, $2, $3, $4, $5)`,
+         VALUES ($1, 'BID_HOLD', -$2::bigint, $2::bigint, $3, $4, $5)`,
         [
           request.user.id,
           additionalHold,
@@ -346,7 +346,7 @@ router.delete(
         `INSERT INTO ledger_entries
           (user_id, entry_type, available_delta, frozen_delta,
            auction_id, bid_id, description)
-         VALUES ($1, 'BID_RELEASE', $2, -$2, $3, $4, 'Puja retirada antes del cierre')`,
+         VALUES ($1, 'BID_RELEASE', $2::bigint, -$2::bigint, $3, $4, 'Puja retirada antes del cierre')`,
         [request.user.id, bid.amount, auction.id, bid.id],
       );
       return { closed: false };
